requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
certifi==2022.6.15
cffi==1.14.6
charset-normalizer==2.1.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
h5py==3.3.0
huepy==1.2.1
idna==3.3
ignite==1.1.0
imageio==2.9.0
importlib-metadata==4.12.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.39.0
MarkupSafe==2.0.1
matplotlib==2.2.4
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numba==0.56.0
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.4.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.1.3
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-ignite==0.4.9
pytz==2021.1
PyYAML==5.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.7.0
seaborn==0.11.2
Send2Trash==1.7.1
setuptools==45.2.0
six==1.16.0
sklearn==0.0
tabulate==0.8.10
terminado==0.10.1
testpath==0.5.0
threadpoolctl==3.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.34.2
widgetsnbextension==3.5.1
wrapt==1.14.1
yacs==0.1.8
zipp==3.8.1